requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
beautifulsoup4==4.14.3
bidict==0.23.1
black==25.12.0
blinker==1.9.0
bs4==0.0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
coverage==7.13.1
cssselect==1.3.0
curl_cffi==0.14.0
cycler==0.12.1
Deprecated==1.3.1
diskcache==5.6.3
dnspython==2.8.0
eventlet==0.40.4
fake-useragent==2.2.0
filelock==3.20.3
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.2
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
librt==0.7.7
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nltk==3.9.2
nodriver==0.48.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
parse==1.20.2
pathspec==1.0.3
patsy==1.0.2
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.1
pyquery==2.0.1
pyreadline3==3.5.4
PySocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-louvain==0.16
python-socketio==5.16.0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-html==0.10.0
rich==14.2.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
statsmodels==0.14.6
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.252
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
vaderSentiment==3.3.2
w3lib==2.3.1
websockets==16.0
Werkzeug==3.1.5
wheel==0.45.1
wrapt==2.0.1
wsproto==1.3.2
yarl==1.22.0
